Refactoring in Series Controller
